+++ app/database/context.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
 
-from app.configuration.environment import DATABASE_URL
+from app.environment import DATABASE_URL
 engine = create_engine(DATABASE_URL)
 
 session = sessionmaker(autocommit=False, autoflush=False, bind=engine)
